requirements.txt
numpy
pandas
matplotlib
scipy
pygame
RPi.GPIO
pyserial
objgraph
psutil
